+++ src/test/java/konkuk/thip/book/adapter/in/web/BookDetailSearchControllerTest.java
@@ -16,7 +16,7 @@ import konkuk.thip.user.adapter.out.persistence.UserRoomJpaRepository;
 import konkuk.thip.feed.adapter.out.jpa.FeedJpaEntity;
 import konkuk.thip.feed.adapter.out.persistence.FeedJpaRepository;
 import konkuk.thip.saved.adapter.out.jpa.SavedBookJpaEntity;
-import konkuk.thip.saved.adapter.out.persistence.SavedBookJpaRepository;
+import konkuk.thip.saved.adapter.out.persistence.repository.SavedBookJpaRepository;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
